Download menu and table button improvement + TS migration
