+++ src/components/Table/Exporter/ExternalExportButton/style.css
@@ -1,3 +1,13 @@
-.browse-btn-api {
-  text-align: center;
+.btn-testing {
+  display: flex;
+  justify-content: left;
+  align-items: center;
+  gap: 10px;
+}
+
+.file-button {
+  display: flex;
+  height: 2.5em;
+  /* to match loading SVG */
+  align-items: center;
 }
