+++ src/pages/Protein/index.js
@@ -200,7 +200,6 @@ class List extends PureComponent /*:: <ListProps> */ {
           >
             <Exporter>
               <div className={f('menu-grid')}>
-                <label htmlFor="fasta">FASTA</label>
                 <AllProteinDownload
                   description={description}
                   iconType="FASTA"
@@ -209,7 +208,6 @@ class List extends PureComponent /*:: <ListProps> */ {
                   fileType="fasta"
                   name="fasta"
                 />
-                <label htmlFor="json">JSON</label>
                 <AllProteinDownload
                   name="json"
                   description={description}
@@ -217,7 +215,6 @@ class List extends PureComponent /*:: <ListProps> */ {
                   count={_payload.count}
                   fileType="json"
                 />
-                <label htmlFor="tsv">TSV</label>
                 <AllProteinDownload
                   name="tsv"
                   description={description}
@@ -225,9 +222,7 @@ class List extends PureComponent /*:: <ListProps> */ {
                   count={_payload.count}
                   fileType="tsv"
                 />
-                <label htmlFor="api">API</label>
                 <ExternalExportButton type={'api'} url={url} />
-                <label htmlFor="code">CODE</label>
                 <ExternalExportButton
                   search={search}
                   type={'scriptgen'}
