+++ src/subPages/SimilarProteins/Table/index.tsx
@@ -108,7 +108,6 @@ const SimilarProteinTable = ({
         <PageSizeSelector />
         <Exporter>
           <div className={css('menu-grid')}>
-            <label htmlFor="fasta">FASTA</label>
             <AllProteinDownload
               description={description}
               ida={ida}
@@ -116,7 +115,6 @@ const SimilarProteinTable = ({
               count={payload.count}
               fileType="fasta"
             />
-            <label htmlFor="json">JSON</label>
             <AllProteinDownload
               description={description}
               ida={ida}
@@ -124,7 +122,6 @@ const SimilarProteinTable = ({
               count={payload.count}
               fileType="json"
             />
-            <label htmlFor="tsv">TSV</label>
             <AllProteinDownload
               description={description}
               ida={ida}
@@ -132,14 +129,11 @@ const SimilarProteinTable = ({
               count={payload.count}
               fileType="tsv"
             />
-            <label htmlFor="api">API</label>
             <ExternalExportButton
               type={'api'}
               url={getAPIURLForSimilarProteins(api, ida, db)}
             />
           </div>
-
-          <label htmlFor="code">CODE</label>
           <ExternalExportButton
             search={search}
             type={'scriptgen'}
